+++ packages/pluggableWidgets-CN/syy-button-web/src/ButtonWeb.tsx
@@ -2,7 +2,8 @@ import { ReactNode, createElement, useCallback } from "react";
 import { ButtonWebContainerProps } from "../typings/ButtonWebProps";
 import { executeAction } from "@mendix/piw-utils-internal";
 import { checkPathPermission } from "./utils/utils";
-import { PopConpemnet, ButtonConpemnet } from "./components/ButtonWeb";
+import { Button } from "./components/Button";
+
 import "./ui/ButtonWeb.css";
 
 import { Modal } from "antd";
@@ -65,39 +66,31 @@ export function ButtonWeb(props: ButtonWebContainerProps): ReactNode {
     }
     const shapeData = shape === "circle" ? "circle" : shape === "round" ? "round" : undefined;
     const textData = text.trim() === "" ? undefined : text;
-
-    const ButtonWebView = (
-        <ButtonConpemnet
-            disabled={!!(disabled && disabled.value)}
-            danger={danger && danger.value}
-            loading={loading && loading.value}
-            ghost={ghost}
-            block={block}
-            href={href ? href : undefined}
-            size={size}
-            shape={shapeData}
-            type={type}
-            onClick={onClick}
-            icon={icon?.value}
-        >
-            {textData}
-        </ButtonConpemnet>
-    );
-
     return (
         <div className={props.class} style={props.style} tabIndex={props.tabIndex}>
-            {openconfirm && confirmType === "pop" ? (
-                <PopConpemnet
-                    confirmtitle={confirmtitle}
-                    confirmokText={confirmokText}
-                    confirmcancelText={confirmcancelText}
-                    confirmokType={confirmokType}
-                >
-                    {ButtonWebView}
-                </PopConpemnet>
-            ) : (
-                ButtonWebView
-            )}
+            <Button
+                disabled={!!(disabled && disabled.value)}
+                danger={danger && danger.value}
+                loading={loading && loading.value}
+                ghost={ghost}
+                block={block}
+                href={href ? href : undefined}
+                size={size}
+                shape={shapeData}
+                type={type}
+                onClick={onClick}
+                icon={icon?.value}
+                openconfirm={openconfirm}
+                confirmType={confirmType}
+                confirmokType={confirmokType}
+                confirmtitle={confirmtitle?.value}
+                confirmokText={confirmokText}
+                confirmcancelText={confirmcancelText}
+                onConfirm={onConfirm}
+                onCancel={onCancel}
+            >
+                {textData}
+            </Button>
         </div>
     );
 }