+++ packages/pluggableWidgets-CN/syy-button-web/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Add e2e test and unit test
+
+### Fixed
+
+-   Fix the bug that the confirm type and confirm events of secondary confirmation is invalid
+
 ## [1.1.3] - 2022-6-21
 
 ### Fixed